chore: moved existing utils to own module
