+++ src/sentence/Sentence.ts
@@ -14,6 +14,12 @@ import {
 /* eslint-enable no-unused-vars */
 
 import defaults from '../defaults';
+import {
+  articleAndPluralize,
+  capitalize,
+  getTotalWeightOfEntries,
+  mapToWeightedEntryArray,
+} from './utils';
 
 export class Sentence {
   #templates: WeightedTemplate[] = [];
@@ -265,33 +271,3 @@ export class Sentence {
     return false;
   }
 }
-
-export const articleAndPluralize = (a_an: boolean, plural: boolean, w: string): string => {
-  return `${a_an ? isVowel(w[0]) ? 'an ' : 'a ' : ''}${w}${plural ? 's' : ''}`;
-};
-
-export const capitalize = (str: string): string => str.replace(/^[']*(\w)/, c => c.toUpperCase());
-
-export const isVowel = (c: string): boolean => ['a', 'e', 'i', 'o', 'u'].includes(c);
-
-export const getTotalWeightOfEntries = (entries: WeightedEntry[]): number => entries.reduce((acc, e) => e ? acc + e.weight : acc, 0);
-
-export const mapToWeightedEntryArray = (entries: [], defaultWeight: number = 1): WeightedEntry[] => {
-  return entries.map<WeightedEntry>(element => {
-    const { entry, weight } = (typeof element === 'object') ? element : {
-      entry: element,
-      weight: 1,
-    };
-    return {
-      entry: entry,
-      weight: weight || defaultWeight,
-    };
-  }).filter((weightedEntry, index, newArray) => {
-    const indexOfDuplicate = newArray.slice(0, index).findIndex(el => el.entry === weightedEntry.entry);
-    if (indexOfDuplicate >= 0) {
-      newArray[indexOfDuplicate].weight += weightedEntry.weight;
-      return false;
-    }
-    return true;
-  });
-};
