+++ src/sentence/utils/index.ts
@@ -1,0 +1,10 @@
+export {
+  getTotalWeightOfEntries,
+  mapToWeightedEntryArray,
+} from './array';
+
+export {
+  articleAndPluralize,
+  capitalize,
+  isVowel,
+} from './format';
